requirements.txt
anyio==3.4.0
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs
Babel==2.9.1
backcall==0.2.0
backoff==1.11.1
bidict==0.21.4
bleach==4.1.0
boto3
botocore
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.4
cloudpickle==2.0.0
contextlib2==21.6.0
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docker==4.4.4
entrypoints==0.3
Flask
Flask-Cors
Flask-SocketIO
flatbuffers==2.0
fonttools==4.29.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
greenlet==1.1.2
hyperopt==0.2.7
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous==2.1.0
Jinja2
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.2
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
kiwisolver==1.3.2
kubernetes
MarkupSafe
matplotlib
matplotlib-inline
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
neural-compressor==1.9.1
notebook==6.4.6
numpy==1.21.4
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
Pint==0.16.1
pluggy==1.0.0
prettytable==3.1.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==8.0.0
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pyflann-py3==0.1.0
Pygments==2.11.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pypng==0.0.21
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.3.7
schema==0.7.5
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sigopt==8.2.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tqdm
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension==3.5.2
xgboost==1.5.1
xmltodict==0.12.0
yq
zipp==3.7.0
zope.event
zope.interface
captum
graphviz
seaborn
pyfunctional
dvc>=2.43.1
dvc[ssh]
onnx==1.10.2
onnxruntime==1.10.0
ffi-navigator
lpips==0.1.4
